requirements.txt
flask==3.1.0
firebase-admin==6.2.0
python-dotenv==1.0.0
flask-cors==4.0.0
gunicorn==21.2.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
Flask-Limiter==3.5.0
Flask-Session==0.8.0
